+++ src/routes/profile.js
@@ -1,0 +1,16 @@
+const express = require("express");
+const profileRouter = express.Router();
+const {userAuth} = require("../middlewares/auth");
+
+profileRouter.get("/profile", userAuth, async (req, res) => {
+  try {
+    const user = req.user;
+
+    // send user profile
+    res.status(200).send(user);
+  } catch (error) {
+    res.status(400).send("ERROR: " + error.message);
+  }
+});
+
+module.exports = profileRouter;
